# API/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CI API",
    "description": "API List for CI APP",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:{port}/api/{apiVersion}",
      "variables": {
        "port": {
          "default": "5002",
          "enum": ["5002"]
        },
        "apiVersion": {
          "default": "v1",
          "enum": ["v1"],
          "description": "Api Version"
        }
      }
    }
  ],
...
  "paths": {
...
    "/customers/secure/wishlist/{productId}": {
      "post": {
        "tags": ["Customer"],
        "summary": "Add Product to the wishlist",
        "description": "",
        "operationId": "addToWishlist",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "Product Uuid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order Placed Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["Customer"],
        "summary": "Remove Product to the wishlist",
        "description": "",
        "operationId": "removeFromWishlist",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "Product Uuid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order Placed Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/customers/secure/cart/placeorder": {
      "post": {
        "tags": ["Customer"],
        "summary": "Place the Order requires customer login",
        "description": "",
        "operationId": "placeOrder",
        "responses": {
          "200": {
            "description": "Order Placed Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "token": {
                "type": "string"
              },
              "user": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string"
                  },
                  "userUuid": {
                    "type": "string"
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "userName": {
                    "type": "string"
                  },
                  "userType": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "dob": {
                    "type": "string",
                    "format": "date"
                  },
                  "gender": {
                    "type": "string",
                    "enum": ["Male", "Female", "Other"]
                  },
                  "phoneNumber": {
                    "type": "string",
                    "pattern": "^[0-9]{10}$"
                  },
                  "password": {
                    "type": "string"
                  },
                  "isActive": {
                    "type": "boolean"
                  },
                  "isDeleted": {
                    "type": "boolean"
                  },
                  "createdAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "updatedAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "__v": {
                    "type": "integer"
                  }
                },
                "required": [
                  "_id",
                  "userUuid",
                  "firstName",
                  "lastName",
                  "userName",
                  "userType",
                  "email",
                  "dob",
                  "gender",
                  "phoneNumber",
                  "password",
                  "isActive",
                  "isDeleted",
                  "createdAt",
                  "updatedAt",
                  "__v"
                ]
              }
            },
            "required": ["token", "user"]
          },
          "message": {
            "type": "string"
          }
        },
        "required": ["success", "data", "message"]
      },
...
    },
...
  }
}